Implement activity detection for paused reminders
When activity is detected and the feature is enabled, a notification
will be pushed indicating reminders are paused. Also adds some minor
changes to reminder forms and parsing to support this new feature Also
adds some minor changes to reminder forms and parsing to support this
new feature Also adds some minor changes to reminder forms and parsing
to support this new feature Also adds some minor changes to reminder
forms and parsing to support this new feature

This commit has some minor bugs that have yet to be tracked down.
Unsure which ones are exactly caused by this commit
- GUI shift when toggling feature in reminder form
-