First iteration of single page UI
Converts the app from a multiple HTML page app to a single page app
while preserving a split file structure. This is accomplished by copying
code from other HTML files into the main HTML files based on what is
needed for radio button navigation.

Some of the code to implement this is fairly sloppy, and improvements
are necessary. Hence, it is the first iteration of this feature. It is
likely that this feature introduces an assortment of bugs, so it will
likely need to be thoroughly tested before being stable.