Setting for paused notif audio & improve notif
Tweaked setting backend system to support select menus and store IDs on
just elements. This is to support a feature to allow users to select
sound for the paused reminder notification. This feature was also tweaked
so that is handled by the reminder list and not the reminders themselves.
Reminders still remember if they had a notification sent, however. Thi
may be subject to change.